+++ lib/precise_metronome.dart
@@ -8,5 +8,6 @@ library precise_metronome;
 export 'src/metronome.dart';
 export 'src/time_signature.dart';
 export 'src/voice.dart';
+export 'src/subdivision.dart';
 export 'src/tap_tempo.dart';
 export 'src/background_config.dart';